2 content factory audio/requirements.txt
google-genai==1.47.0
python-dotenv
requests
pathlib
